code/reporter/requirements.txt
slack_sdk
pandas
snowflake-connector-python[pandas]